feat: initialize node status plugin
